requirements.txt
Flask==3.0.3
line-bot-sdk==2.4.2
python-dotenv==1.0.1
deep-translator==1.11.4
langdetect==1.0.9
gunicorn==21.2.0
requests==2.28.2